+++ packages/user/src/index.ts
@@ -12,6 +12,9 @@ declare module "@dzangolab/fastify-config" {
   interface ApiConfig {
     user: {
       supertokens: SupertokensConfig;
+      features?: {
+        signUp?: boolean;
+      };
     };
   }
 }
@@ -24,8 +27,6 @@ export type {
   UserProfileUpdateInput,
   User,
 } from "./types";
-
-export { default as buildAuthContext } from "./buildContext";
 
 export { default as userProfileResolver } from "./model/user-profiles/resolver";
 export { default as UserProfileService } from "./model/user-profiles/service";