+++ packages/user/src/model/users/resolver.ts
@@ -13,7 +13,7 @@ const Mutation = {
     },
     context: MercuriusContext
   ) => {
-    const service = new Service();
+    const service = new Service(context.config, context.database);
 
     try {
       if (context.user?.id) {
@@ -32,15 +32,38 @@ const Mutation = {
       }
     } catch (error) {
       // FIXME [OP 28 SEP 2022]
-      console.log(error);
+      context.app.log.error(error);
 
       const mercuriusError = new mercurius.ErrorWithProps(
         "Oops, Something went wrong"
       );
       mercuriusError.statusCode = 500;
+
       return mercuriusError;
     }
   },
 };
 
-export default { Mutation };
+const Query = {
+  me: async (
+    parent: unknown,
+    arguments_: unknown,
+    context: MercuriusContext
+  ) => {
+    const service = new Service(context.config, context.database);
+    if (context.user?.id) {
+      return service.getUserById(context.user.id);
+    } else {
+      context.app.log.error("Cound not get user id from mercurius context");
+
+      const mercuriusError = new mercurius.ErrorWithProps(
+        "Oops, Something went wrong"
+      );
+      mercuriusError.statusCode = 500;
+
+      return mercuriusError;
+    }
+  },
+};
+
+export default { Mutation, Query };