+++ packages/user/src/model/users/service.ts
@@ -1,7 +1,26 @@
 import Session from "supertokens-node/recipe/session";
 import ThirdPartyEmailPassword from "supertokens-node/recipe/thirdpartyemailpassword";
+import UserRoles from "supertokens-node/recipe/userroles";
+
+import userProfileService from "../user-profiles/service";
+
+import type {
+  UserProfile,
+  UserProfileCreateInput,
+  UserProfileUpdateInput,
+} from "../../types";
+import type { ApiConfig } from "@dzangolab/fastify-config";
+import type { Database } from "@dzangolab/fastify-slonik";
+import type { QueryResultRow } from "slonik";
 
 class UserService {
+  config: ApiConfig;
+  database: Database;
+  constructor(config: ApiConfig, database: Database) {
+    this.config = config;
+    this.database = database;
+  }
+
   changePassword = async (
     userId: string,
     oldPassword: string,
@@ -77,6 +96,39 @@ class UserService {
         message: "Password cannot be empty",
       };
     }
+  };
+
+  getUserById = async (userId: string) => {
+    const user = await ThirdPartyEmailPassword.getUserById(userId);
+
+    const service: userProfileService<
+      UserProfile & QueryResultRow,
+      UserProfileCreateInput,
+      UserProfileUpdateInput
+    > = new userProfileService(this.config, this.database);
+
+    /* eslint-disable-next-line unicorn/no-null */
+    let profile: UserProfile | null = null;
+
+    try {
+      profile = await service.findById(userId);
+    } catch {
+      // FIXME [OP 2022-AUG-22] Handle error properly
+      // DataIntegrityError
+    }
+
+    const roles = await UserRoles.getRolesForUser(userId);
+
+    return {
+      status: "OK",
+      user: {
+        email: user?.email,
+        id: userId,
+        profile: profile,
+        roles: roles.roles,
+        timeJoined: user?.timeJoined,
+      },
+    };
   };
 }